Add a test server hosting the same app using warp and hope.
Using the wai-app-static file server app.
We use a Middleware to add SPDY advertisement (in the request headers) for
each page request to the warp server.
Unfortunately it seems Chrome 16 and Chrome 17 doesn't implement the
spdy/2 alternate protocol, but only npn-spdy/2. This means we won't be able
to properly advertise the SPDY to Chrome until we have support for NPN in
TLS. I have not tested Firefox 11.