Start generalized API
Add new modules to the library, which mostly are independant to the existing
modules. This is a new start of the library, as an attempt to provide a
generalized API, which should be useable to build both clients and servers.
It does not build on a existing web framework.
Currently uses io-streams and tls.